fixed buffer-paging issue in putsource()
This patch fixes a bug where the block number and the block index would get out
of sync. Inside putsource() there was code that boiled down to this:

	do {
		if(...) {
			Change = YES;
			cp = blockp;
		}

		putline(output);
	} while (blockp != NULL && getrefchar() != '\n');
	if (Change == YES) blockp = cp;

It was possible for putline() to read in a new block. The last line could then
reset blockp WITHOUT reloading the old block. This effectively skipped a block's
worth of data resulting in a cscope report with some extra bogus data and some
missing valid data