Fixed a bug tickled by the previous commit
putsource() was calling putline(output) and then resetting retreat to NO without
checking to make sure retreat IS now NO (i.e. whether we're now in the same
block we were before). Here we get rid of retreat entirely and actually check
for the thing we care about: which block we are in.