*cscope* buffer now works a bit like a diff-mode buffer
One can now navigate through results in a search with M-N,M-P and kill a result
with M-K.

In a later patch one will be able to now navigate through FILES in a search with
M-n,M-p and kill a file with M-k.

This is like files and hunks in a diff respectively. The 'v' and 'V' bindings
have been removed for consistency. These bindings all work from within a
*cscope* buffer only. There is NO corresponding C-c s ... binding